refactor(security): fix nft slice
