+++ src/Components/Providers/EncryptedStorageProvider/ApplicationSecurityProvider.tsx
@@ -3,7 +3,7 @@ import { MMKV } from "react-native-mmkv"
 import { debug, error, HexUtils, warn } from "~Utils"
 import { BiometricState, SecurityLevelType, WALLET_STATUS } from "~Model"
 import {
-    PreviousIntallation,
+    PreviousInstallation,
     SecurityConfig,
     SecurityUpgradeBackup,
     StorageEncryptionKeyHelper,
@@ -202,7 +202,7 @@ export const ApplicationSecurityProvider = ({
             if (encryptedStorageKeys.length === 0) {
                 warn("No keys found in encrypted storage, user is onboarding")
 
-                await PreviousIntallation.clearOldStorage()
+                await PreviousInstallation.clearOldStorage()
 
                 OnboardingStorage.getAllKeys().forEach(key => {
                     OnboardingStorage.delete(key)
