+++ src/Components/Providers/EncryptedStorageProvider/Helpers/index.ts
@@ -2,7 +2,7 @@ import WalletEncryptionKeyHelper from "./WalletEncryptionKeyHelper"
 import StorageEncryptionKeyHelper from "./StorageEncryptionKeyHelper"
 import SecurityConfig from "./SecurityConfig"
 import SecurityUpgradeBackup from "./SecurityUpgradeBackup"
-import PreviousIntallation from "./PreviousIntallation"
+import PreviousInstallation from "./PreviousInstallation"
 
 export * from "./Onboarding"
 
@@ -11,5 +11,5 @@ export {
     StorageEncryptionKeyHelper,
     SecurityConfig,
     SecurityUpgradeBackup,
-    PreviousIntallation,
+    PreviousInstallation,
 }
